+++ differential-dataflow/src/trace/cursor/wrappers/mod.rs
@@ -7,3 +7,4 @@
 //! would supply its own module here, and reuse the same time rules.
 
 pub mod enter;
+pub mod frontier;
